Updated the mod to Minecraft version 1.21.11 and hopefully fixed all the resulting problems.
Rendering has been split up into phases:
World Extraction: All the rendering steps for the world are constructed and buffered.
World Rendering: The buffered work is handed off to another thread which dispatches the rendering on the GPU.
GUI Extraction: In parallel to the World rendering the GUI is extracted.
GUI Rendering: The Gui is dispatched for render on the GPU.

+++ src/main/kotlin/floppacoding/mithras/utils/GeometryHelper.kt
@@ -49,6 +49,6 @@ object GeometryHelper {
         }
     }
 
-    private const val DEG_TO_RAD: Float = Math.PI.toFloat() / 180f
+    private const val DEG_TO_RAD: Double = Math.PI / 180
     private const val RAD_TO_DEG: Float = 180f / Math.PI.toFloat()
 }
